tests/wpt/web-platform-tests/webmessaging/with-options/broken-origin.html

<!doctype html>
<title>resolving broken url</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws('SYNTAX_ERR', function() {
    postMessage('', {targetOrigin: 'http://foo bar'});
  }, 'should have failed to resolve');
});
</script>
